dont verify just decode
